move DealMeta to its own file and extract an expiryFormatter prop for easier testing
